Benchmarking times!
We benchmark times on your machine vs. typical times from Qat against the Broda list. While not apples to apples, we really shouldn't fall (say) an order of magnitude behind Qat, no matter what machine we're on.  On my machine we're beating the tar out of Qat on everything except the one I've marked as slow.